+++ xeon_phi/scalar/mul/mul.c
@@ -38,7 +38,7 @@ int main (int argc, char *argv[]) {
     //==================================================================
     // Benchmark variables
     uint32_t th_id = 0;
-    uint32_t i = 0;
+    uint64_t i = 0;
     uint32_t error_count = 0;
     uint32_t final_ref_word = pow(ref_word, 32);
