+++ xeon_phi/scalar/rol/rol.c
@@ -39,7 +39,7 @@ int main (int argc, char *argv[]) {
     //==================================================================
     // Benchmark variables
     uint32_t th_id = 0;
-    uint32_t i = 0;
+    uint64_t i = 0;
     uint32_t error_count = 0;
 
     #pragma offload target(mic)
