+++ xeon_phi/vector/or/or.c
@@ -41,7 +41,7 @@ int main(int argc, char *argv[]) {
     // Benchmark variables
     double startTime,  duration;
     uint32_t th_id = 0;
-    uint32_t i = 0;
+    uint64_t i = 0;
     uint32_t error_count = 0;
     uint32_t zeros;
